Move redux store connection to AppContainer, and convert containers to components that receive props from it
