+++ src/view/Table/index.js
@@ -1,9 +1,7 @@
 // Barrel file to export all Table components
 
-import TableContainer from "./TableContainer";
 import TableComponent from "./TableComponent";
 import TableRowComponent from "./TableRowComponent";
 import TableCellComponent from "./TableCellComponent";
 
-export default TableContainer;
 export { TableComponent, TableRowComponent, TableCellComponent };
